llm-challenge/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
passlib[bcrypt]
PyJWT
requests
python-dotenv
python-multipart
groq
psycopg2-binary
bcrypt==3.2.2
streamlit
requests
qdrant_client
psycopg2-binary
pydantic[email]
sentence-transformers